requirements.txt
nltk==3.3
spacy==3.0.6
language_check==1.1
pandas>=0.23.0
numpy>=1.19.3
textblob==0.15.3
pyspellchecker==0.5.0
gensim==3.4.0
Flask==0.10.1
gunicorn==20.1.0
itsdangerous==0.24
Jinja2==2.10
tensorflow-cpu>=2.3.1
Werkzeug==0.14.1
MarkupSafe==0.23